pyproject.toml
[tool.poetry]
name = "Narratave"
version = "0.1.0"
description = ""
authors = ["Suvodip Mondal <smondal229@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.24.1"
psycopg2 = "2.9.6"
django-environ = "^0.10.0"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
django = "4.1.3"
uvicorn = "0.23.1"
firebase-admin = "^6.2.0"
pillow = "10.0.0"
pydub = "0.25.1"
ffmpeg-python = "^0.2.0"
serpy = "^0.3.1"
psycopg2-binary = "2.9.6"
django-extensions = "^3.2.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
